Added DrmDefaults to factor out default values common to all Drm systems
