+++ src/main/scala/loamstream/drm/uger/UgerDefaults.scala
@@ -6,20 +6,13 @@ import loamstream.model.quantities.Cpus
 import loamstream.model.quantities.Memory
 import java.nio.file.Paths
 import java.nio.file.Path
+import loamstream.drm.DrmDefaults
 
 /**
  * @author clint
  * Oct 11, 2017
  */
-object UgerDefaults {
-  val maxNumJobsPerTaskArray: Int = 2000 //scalastyle:ignore magic.number
-  
-  val cores: Cpus = Cpus(1)
-
-  val memoryPerCore: Memory = Memory.inGb(1)
-    
-  val maxRunTime: CpuTime = CpuTime.inHours(2)
-  
+object UgerDefaults extends DrmDefaults {
   val queue: Queue = Queue("broad")
   
   val extraPathDir: Path = Paths.get("/humgen/diabetes/users/dig/miniconda2/bin")
@@ -27,8 +20,6 @@ object UgerDefaults {
   val condaEnvName: String = "loamstream_v1.0"
   
   val staticJobSubmissionParams: String = "-cwd -shell y -b n"
-  
-  val maxRetries: Int = 9
   
   val maxQacctCacheSize: Int = 100
 }
